+++ src/kivecli/download.py
@@ -14,7 +14,7 @@ from .collect_run_files import collect_run_files
 from .dataset import Dataset
 from .escape import escape
 from .usererror import UserError
-from .await_containerrrun import await_containerrun
+from .await_containerrun import await_containerrun
 from .runfilesfilter import RunFilesFilter
 from .kiverun import KiveRun
