+++ src/kivecli/runkive.py
@@ -19,7 +19,7 @@ from .parsecli import parse_cli
 from .login import login
 from .url import URL
 from .escape import escape
-from .await_containerrrun import await_containerrun
+from .await_containerrun import await_containerrun
 from .runfilesfilter import RunFilesFilter
 from .findbatches import findbatches
 from .kiverun import KiveRun
