+++ src/kivecli/watch.py
@@ -8,7 +8,7 @@ from .mainwrap import mainwrap
 from .parsecli import parse_cli
 from .login import login
 from .findrun import find_run
-from .await_containerrrun import await_containerrun
+from .await_containerrun import await_containerrun
 
 
 def cli_parser() -> argparse.ArgumentParser:
